CS-3108: Moves wyre custodial transfer into job task so that we can retry on failure (#2680)
* cleanup unnecessary types

* Moves wyre custodial transfer into job task so that we can retry on failure

* fix broken test

* added logging
